+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - **Breaking**: Updated minimum Dart SDK version to 2.19.0.
 - **Breaking**: Renamed `HolodexClient` constructor parameter `httpClient` to `client`.
+- **Breaking**: Renamed `HolodexClient.getVideosRelatedToChannel` to `getChannelRelatedVideos`.
 - **Breaking**: Renamed `HolodexClient.getVideoMetadata` parameter `timestampComments` to `includeTimestampComments`.
 - **Breaking**: Renamed `HolodexClient.getVideoMetadata` parameter `recommendationLanguages` to `languages`.
 - **Breaking**: Renamed `client.dart` to `holodex_client.dart`. Technically a breaking change, but only if